Save extended info about files to storage

+++ src/main/kotlin/com/github/pberdnik/dependenciesanalyzerplugin/old/graph/DependencyGraph.kt
@@ -55,7 +55,7 @@ class DependencyGraph {
         topSorted.forEach { node ->
             val onlyRed = node.onlyRed
             if (onlyRed != null) {
-                redNodes[onlyRed.id] = (redNodes[onlyRed.id] ?: 0) + node.codeFile.size
+                redNodes[onlyRed.id] = (redNodes[onlyRed.id] ?: 0) + node.codeFile.size.toInt()
             }
         }
     }
